Refactored scripts and utils
