+++ Code/monet_cyclegan/__init__.py
@@ -2,3 +2,4 @@ from . import consts
 from . import data_acquisition
 from . import deployment
 from . import modeling
+from . import scripts
