+++ Code/monet_cyclegan/data_acquisition/__init__.py
@@ -1,3 +1,2 @@
 from . import datasets
-from . import utils
 from . import download_dataset
