+++ Code/monet_cyclegan/data_acquisition/download_dataset.py
@@ -1,9 +1,8 @@
 import os
-from argparse import ArgumentParser
 from pathlib import Path
 
-from .utils import download_extract_zip
 from ..consts import KAGGLE_DATASET_PATH, KAGGLE_DATASET_URL
+from ..utils import download_extract_zip
 
 
 def download_kaggle_dataset(url: str = KAGGLE_DATASET_URL,
@@ -21,14 +20,3 @@ def download_kaggle_dataset(url: str = KAGGLE_DATASET_URL,
         raise OSError(f'Directory "{output_dir}" is not empty.')
 
     download_extract_zip(url, output_dir)
-
-
-if __name__ == '__main__':
-    parser = ArgumentParser()
-    parser.add_argument('--dir', type=str, default=PHOTO_TFREC_DIR)
-    parser.add_argument('--ext', type=str, default='tfrec')
-    parser.add_argument('--file', '-f', type=str)
-    parser.add_argument('--output', '-o', type=str, default=OUTPUT_DIR)
-    args = parser.parse_args()
-
-    download_kaggle_dataset()
