+++ Code/monet_cyclegan/deployment/generate_images.py
@@ -2,16 +2,11 @@
 Generates images from the trained model.
 """
 import os
-from argparse import ArgumentParser
 
-from numpy import uint8
-
-from .load_model import load_cyclegan_model
-from .utils import translate_image, save_image, make_directory, tensor_to_image
-from ..consts import IMAGE_SIZE, CHANNELS, PHOTO_TFREC_DIR, MONET_GENERATOR_WEIGHT_PATH, WEIGHT_OUTPUT_DIR, OUTPUT_DIR, \
-    PHOTO_GENERATOR_WEIGHT_PATH, EPOCHS
-from ..data_acquisition.utils import read_image, read_tfrecorddataset, get_filenames
+from ..consts import IMAGE_SIZE, CHANNELS
 from ..modeling.model import CycleGan
+from ..modeling.predict import translate_image
+from ..utils import read_image, read_tfrecorddataset, get_filenames, save_image, make_directory, tensor_to_image
 
 
 def generate_image(cyclegan_model: CycleGan, input_path: str, output_dir: str) -> None:
@@ -73,8 +68,6 @@ def generate_images(cyclegan_model: CycleGan,
 
         photos = photos.batch(1)
 
-        make_directory(output_dir)
-
         for i, image in enumerate(photos):
             generated_image = translate_image(cyclegan_model=cyclegan_model,
                                               image=image)
@@ -93,40 +86,3 @@ def generate_images(cyclegan_model: CycleGan,
             generate_image(cyclegan_model=cyclegan_model,
                            input_path=filename,
                            output_dir=output_dir)
-
-
-def main() -> None:
-    parser = ArgumentParser()
-    parser.add_argument('--input', type=str, default=PHOTO_TFREC_DIR)
-    parser.add_argument('--output', '-o', type=str, default=OUTPUT_DIR)
-    parser.add_argument('--file', '-f', type=str)
-    parser.add_argument('--ext', type=str, default='tfrec')
-    parser.add_argument('--num', '-n', type=int, default=None)
-    parser.add_argument('--randomize', action='store_true')
-    parser.add_argument('--no-randomize', dest='randomize', action='store_false')
-    parser.set_defaults(randomize=False)
-    parser.add_argument('--with-original', action='store_true')
-    parser.add_argument('--without-original', dest='with-original', action='store_false')
-    parser.set_defaults(with_original=False)
-    args = parser.parse_args()
-
-    model = load_cyclegan_model(
-        monet_generator_weights_path=f'{WEIGHT_OUTPUT_DIR}/epoch{EPOCHS}/{MONET_GENERATOR_WEIGHT_PATH}',
-        photo_generator_weights_path=f'{WEIGHT_OUTPUT_DIR}/epoch{EPOCHS}/{PHOTO_GENERATOR_WEIGHT_PATH}')
-
-    if args.file:
-        generate_image(cyclegan_model=model,
-                       input_path=args.file,
-                       output_dir='.')
-    else:
-        generate_images(cyclegan_model=model,
-                        input_dir=args.input,
-                        input_ext=args.ext,
-                        output_dir=args.output,
-                        sample_size=args.num,
-                        randomize=args.randomize,
-                        with_original=args.with_original)
-
-
-if __name__ == '__main__':
-    main()
