+++ Code/monet_cyclegan/scripts/download_dataset.py
@@ -1,0 +1,9 @@
+from ..data_acquisition.download_dataset import download_kaggle_dataset
+
+
+def main():
+    download_kaggle_dataset()
+
+
+if __name__ == '__main__':
+    main()
